Create DCP_FlyWay_20221117
First, let us identify the digits that can be part of our number. 0, 1, and 8 all have rotational symmetry, while 6 and 9 complement each other's rotational symmetry. What this means is that, for example, if a 6 begins a number, a 9 must come at the end. We can more neatly describe this state of affairs by providing a map from each digit to their flipped image:

To form each possible N-digit number, we can use backtracking. We start with one key, say 1. To this we can add any of the five keys as the second digit. We continue adding in this way until our string reaches a length of half the required length.

What happens next depends on whether N is even or odd. If N is even, we need only add the mirror images of each digit to the end of the string, in reverse order. For example, suppose we are finding all six-digit numbers, and our current string is 186. The mirror images of these digits are [1, 8, 9], and so our strobogrammatic result is 186981.

On the other hand, if N is odd, we should not flip and reverse the last digit, as this will be the center of our number. This implies that we cannot allow 6 or 9 to be the final digit of our candidate string. For example, 180 can safely become 18081, but we must discard 186.

Once we generate all symmetric number starting with 1, we proceed similarly through the other keys. Along the way, we append each strobogrammatic number found to a result list.